# api/openapi.yaml
openapi: 3.0.3
info:
  title: Muasal API
  version: 0.1.0
  description: >-
    The API the web app uses, open to scripts through personal access tokens (FSD §14.3). Send
    `Authorization: Bearer msl_…`; a token acts with its owner's visibility and role, a read-only token only reads
    (and may ask), and each token takes 60 requests a minute, bursting to 120. Browser sessions use the sid cookie
    and must send the app's Origin on writes.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
paths:
...
  /projects/{key}/labels:
    parameters:
      - { name: key, in: path, required: true, schema: { type: string } }
    get:
      operationId: listProjectLabels
      tags: [tickets]
      description: The labels the project's tickets use, most used first (MSL-56).
      responses:
        "200":
          description: Labels with how many tickets carry each.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [label, uses]
                      properties:
                        label: { type: string }
                        uses: { type: integer }
        default: { $ref: "#/components/responses/Problem" }
...
  /projects/{key}/tickets:
    parameters:
      - { name: key, in: path, required: true, schema: { type: string } }
    get:
      operationId: listTickets
      tags: [tickets]
      description: The project's tickets the caller may see, filtered and sorted; pages follow next_cursor.
      parameters:
        - { name: status_id, in: query, schema: { type: integer, format: int64 } }
        - { name: category, in: query, schema: { $ref: "#/components/schemas/StatusCategory" } }
        - { name: open, in: query, schema: { type: boolean }, description: Only To do and In progress statuses. }
        - { name: closed_days, in: query, schema: { type: integer, format: int32, minimum: 1 }, description: Closed tickets only when closed within this many days; the board asks for 14. }
        - { name: type, in: query, schema: { $ref: "#/components/schemas/TicketType" } }
        - { name: label, in: query, schema: { type: string, maxLength: 30 }, description: Tickets with this label (MSL-56). }
        - { name: accepted, in: query, schema: { type: boolean }, description: "Tickets the client accepted (true) or not yet (false) (MSL-66)." }
        - { name: release_id, in: query, schema: { type: integer, format: int64 }, description: Tickets of this release (MSL-67). }
        - { name: client_id, in: query, schema: { type: integer, format: int64 } }
        - { name: core, in: query, schema: { type: boolean }, description: Only core work (no client). }
        - { name: assignee_id, in: query, schema: { type: integer, format: int64 } }
        - { name: mine, in: query, schema: { type: boolean }, description: Only tickets assigned to the caller. }
        - { name: unassigned, in: query, schema: { type: boolean }, description: Only tickets nobody owns. }
        - name: due
          in: query
          schema: { type: string, enum: [overdue, week] }
          description: Open tickets past their due date (overdue), or due from today through 7 days ahead (week), as Home counts them. Today is the date in the caller's timezone.
        - name: stale_days
          in: query
          schema: { type: integer, format: int32, minimum: 1, maximum: 365 }
          description: Open tickets unchanged for at least this many days; comments, files and decision records count as changes.
        - { name: node_id, in: query, schema: { type: integer, format: int64 }, description: Tickets on this node or its sub-nodes. }
        - { name: q, in: query, schema: { type: string, maxLength: 200 }, description: Words in the title, or a ticket key. }
        - { name: missing, in: query, schema: { type: string, enum: [reason, menus, weak_reason] }, description: "weak_reason: a reason under 20 characters once stock phrases such as 'permintaan klien' are removed, the form's hint rule (R-DC-8)." }
        - { name: sort, in: query, schema: { type: string, enum: [updated, created, key, priority, due] } }
        - { name: limit, in: query, schema: { type: integer, format: int32, minimum: 1, maximum: 1000 } }
        - { name: cursor, in: query, schema: { type: string } }
        - name: format
          in: query
          schema: { type: string, enum: [json, csv] }
          description: csv downloads every ticket of the filter, up to 10,000, as UTF-8 CSV instead of a page (FSD §8.5).
      responses:
        "200":
          description: One page of tickets; with format=csv, the whole filter as text/csv instead.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketPage" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: createTicket
      tags: [tickets]
      description: Members and project admins. The ticket starts in the default status unless an open status_id is given.
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string, maxLength: 200 }
          description: >-
            A retry with the same key within 24 hours returns the ticket the first request created (200, with
            Idempotent-Replayed true) instead of a second ticket (FSD §17.1).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketCreate" }
      responses:
        "201":
          description: The new ticket; ETag carries its version.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  responses:
    Problem:
      description: An error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
...
    Ref:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
...
    StatusCategory:
      type: string
      enum: [todo, in_progress, done, cancelled]
    Status:
      type: object
      required: [id, name, category, color, position, is_default]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        category: { $ref: "#/components/schemas/StatusCategory" }
        color: { type: string, example: "#2563EB" }
        position: { type: integer, format: int32 }
        is_default: { type: boolean }
...
    TicketType:
      type: string
      enum: [bug, change_request, feature]
    Priority:
      type: string
      enum: [low, medium, high, urgent]
    TicketRequester:
      type: object
      required: [kind, id, name, title]
      properties:
        kind: { type: string, enum: [contact, user] }
        id: { type: integer, format: int64 }
        name: { type: string }
        title: { type: string, nullable: true, example: HR Manager }
    NodeRef:
      type: object
      required: [id, name, archived]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        archived: { type: boolean }
    Attachment:
      type: object
      required: [id, filename, content_type, size_bytes, uploader, created_at]
      properties:
        id: { type: integer, format: int64 }
        filename: { type: string }
        content_type: { type: string }
        size_bytes: { type: integer, format: int64 }
        uploader: { $ref: "#/components/schemas/Ref" }
        created_at: { type: string, format: date-time }
    Ticket:
      type: object
      required: [id, key, project_key, title, type, description, reason, status, requester, reporter, priority, due_date, nodes, attachments, links, version, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        key: { type: string, example: HRIS-231 }
        project_key: { type: string }
        title: { type: string }
        type: { $ref: "#/components/schemas/TicketType" }
        description: { type: string }
        reason: { type: string }
        status: { $ref: "#/components/schemas/Status" }
        client: { $ref: "#/components/schemas/Ref" }
        requester: { $ref: "#/components/schemas/TicketRequester" }
        reporter: { $ref: "#/components/schemas/Ref" }
        assignee: { $ref: "#/components/schemas/Ref" }
        priority: { $ref: "#/components/schemas/Priority" }
        due_date: { type: string, format: date, nullable: true }
        estimate_hours: { type: number, format: double, nullable: true, description: The effort estimate in hours (MSL-54). }
        labels: { type: array, items: { type: string }, description: "Free-text labels, lowercased (MSL-56)." }
        following: { type: boolean, description: The caller follows the ticket (MSL-57). }
        acceptance: { $ref: "#/components/schemas/TicketAcceptance" }
        release: { $ref: "#/components/schemas/Ref" }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/NodeRef" }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        version: { type: integer, format: int32 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        closed_at: { type: string, format: date-time, nullable: true, description: Set on close, cleared on reopen. }
        decision: { $ref: "#/components/schemas/DecisionRecord" }
        links:
          type: array
          description: Links to tickets the reader can see, oldest first (FSD §8.8).
          items: { $ref: "#/components/schemas/TicketLink" }
        code: { $ref: "#/components/schemas/TicketCode" }
    TicketCreate:
      type: object
      required: [type, title, node_ids]
      properties:
        type: { $ref: "#/components/schemas/TicketType" }
        title: { type: string, maxLength: 200 }
        client_id: { type: integer, format: int64, description: Omitted for core work (all clients). }
        requester_contact_id: { type: integer, format: int64 }
        requester_user_id: { type: integer, format: int64, description: Without a requester the reporter is the requester. }
        node_ids:
          type: array
          items: { type: integer, format: int64 }
        reason: { type: string, maxLength: 2000 }
        description: { type: string, maxLength: 50000 }
        assignee_id: { type: integer, format: int64 }
        priority: { $ref: "#/components/schemas/Priority" }
        due_date: { type: string, format: date }
        estimate_hours: { type: number, format: double, minimum: 0, maximum: 9999, description: Hours of effort; omitted means no estimate (MSL-54). }
        labels: { type: array, maxItems: 10, items: { type: string, maxLength: 30 }, description: "Up to 10 labels of 1 to 30 characters; omitted means none (MSL-56)." }
        release_id: { type: integer, format: int64, description: "A release of this project; omitted means none (MSL-67)." }
        status_id: { type: integer, format: int64, description: An open status; omitted means the project's default. }
        note_key: { type: string, description: "A decision note of this project the caller may see, e.g. HRIS-DN7: the ticket joins its tickets, as when filed from one of its action items (MSL-11)." }
...
    DecisionOutcome:
      type: string
      enum: [implemented, rejected]
    DecisionState:
      type: string
      enum: [draft, confirmed]
    DecisionRecord:
      type: object
      required: [what_changed, why, alternatives, outcome, state]
      properties:
        what_changed: { type: string }
        why: { type: string }
        alternatives: { type: string }
        outcome: { $ref: "#/components/schemas/DecisionOutcome" }
        state: { $ref: "#/components/schemas/DecisionState" }
        confirmed_by: { $ref: "#/components/schemas/Ref" }
        confirmed_at: { type: string, format: date-time }
        superseded_by: { type: string, description: The key of the ticket that reverses this decision (R-TK-5). }
        ai_drafted: { type: boolean }
    LinkType:
      type: string
      enum: [reverses, extends, related_to]
...
    TicketLink:
      type: object
      required: [id, type, outgoing, ticket]
      properties:
        id: { type: integer, format: int64 }
        type: { $ref: "#/components/schemas/LinkType" }
        outgoing: { type: boolean, description: True when this ticket is the one that reverses or extends; false reads "Reversed by". }
        ticket: { $ref: "#/components/schemas/LinkedTicket" }
    LinkedTicket:
      type: object
      required: [key, title, status]
      properties:
        key: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/Status" }
        closed_at: { type: string, format: date-time, nullable: true }
...
    TicketAcceptance:
      type: object
      description: Who at the client accepted the ticket's work, as in UAT sign-off, and when (MSL-66).
      required: [contact, accepted_on, note]
      properties:
        contact: { $ref: "#/components/schemas/Ref" }
        accepted_on: { type: string, format: date }
        note: { type: string }
...
    TicketSummary:
      type: object
      required: [id, key, title, type, priority, due_date, status_id, requester_name, node_names, missing_reason, updated_at]
      properties:
        id: { type: integer, format: int64 }
        key: { type: string }
        title: { type: string }
        type: { $ref: "#/components/schemas/TicketType" }
        priority: { $ref: "#/components/schemas/Priority" }
        due_date: { type: string, format: date, nullable: true }
        status_id: { type: integer, format: int64 }
        client: { $ref: "#/components/schemas/Ref" }
        assignee: { $ref: "#/components/schemas/Ref" }
        requester_name: { type: string }
        node_names:
          type: array
          items: { type: string }
        missing_reason: { type: boolean }
        labels: { type: array, items: { type: string } }
        accepted_on: { type: string, format: date, description: When the client accepted it (MSL-66). }
        release: { type: string, description: Its release's name (MSL-67). }
        checklist:
          type: object
          description: The description's task list, when it has one (MSL-55).
          required: [done, total]
          properties:
            done: { type: integer }
            total: { type: integer }
        updated_at: { type: string, format: date-time }
    TicketPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TicketSummary" }
        next_cursor: { type: string, nullable: true }
...
    TicketCode:
      type: object
      required: [merge_requests, commits]
      properties:
        merge_requests: { type: array, items: { $ref: "#/components/schemas/CodeMergeRequest" } }
        commits: { type: array, items: { $ref: "#/components/schemas/CodeCommit" } }
    CodeMergeRequest:
      type: object
      required: [repo, number, title, state]
      properties:
        repo: { type: string }
        number: { type: integer }
        title: { type: string }
        state: { type: string }
        merged_at: { type: string, format: date-time, nullable: true }
        url: { type: string, nullable: true }
    CodeCommit:
      type: object
      required: [repo, sha, message]
      properties:
        repo: { type: string }
        sha: { type: string }
        message: { type: string, description: The first line. }
        author: { type: string, nullable: true }
        committed_at: { type: string, format: date-time, nullable: true }
        url: { type: string, nullable: true }
...